backend/pyproject.toml
[project]
name = "networth-tracker"
version = "0.1.0"
description = "A Django REST API for tracking net worth across multiple asset types"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.9,<6.0",
    "djangorestframework>=3.15.0",
    "django-cors-headers>=4.3.0",
    "python-dotenv>=1.0.0",
    "djangorestframework-simplejwt>=5.3.0",
    "requests>=2.32.5",
    "dj-database-url>=2.2.0",
    "psycopg2-binary>=2.9.9",
    "yfinance>=0.2.40",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-django>=4.9.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "pre-commit>=4.0.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "networth_tracker.settings"
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 79
target-version = "py313"
extend-exclude = [
    "migrations",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by formatter
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "lf"

[tool.ruff.lint.isort]
known-first-party = ["networth_tracker"]
